+++ tests/AdminConsole.Tests/Components/Shared/Layouts/NavMenuTests.cs
@@ -67,78 +67,4 @@ public class NavMenuTests : TestContext
         // Assert
         Assert.True(cut.Find("nav").Children.Any());
     }
-
-    [Fact]
-    public void NavMenu_Renders_ExpectedApplicationLinks_ApplicationIsPendingDeletion()
-    {
-        // Arrange
-        var organizationFeatures = _fixture
-            .Build<OrganizationFeaturesContext>()
-            .With(x => x.EventLoggingIsEnabled, true)
-            .Create();
-        _currentContext.SetupGet(x => x.OrganizationFeatures).Returns(organizationFeatures);
-        var features = _fixture
-            .Build<ApplicationFeatureContext>()
-            .With(x => x.EventLoggingIsEnabled, true)
-            .Create();
-        _currentContext.SetupGet(x => x.Features).Returns(features);
-
-        var organization = _fixture.Build<Organization>()
-            .Without(x => x.Applications)
-            .Without(x => x.Admins)
-            .Create();
-        _currentContext.SetupGet(x => x.Organization).Returns(organization);
-        _currentContext.SetupGet(x => x.IsPendingDelete).Returns(true);
-        _currentContext.SetupGet(x => x.InAppContext).Returns(true);
-        _authorizationContext.SetAuthorized("jonas");
-
-        // Act
-        var cut = RenderComponent<NavMenu>();
-
-        // Assert
-        var submenu = cut.Find("div[id=\"app-submenu\"]");
-        var links = submenu.Children.Where(x => x.ClassName == "nav-link").ToImmutableList();
-        Assert.Equal(2, links.Count);
-        Assert.Contains("Settings", links[0].TextContent);
-        Assert.Contains("App Logs", links[1].TextContent);
-    }
-
-    [Fact]
-    public void NavMenu_Renders_ExpectedApplicationLinks_ApplicationIsNotPendingDeletion()
-    {
-        // Arrange
-        var organizationFeatures = _fixture
-            .Build<OrganizationFeaturesContext>()
-            .With(x => x.EventLoggingIsEnabled, true)
-            .Create();
-        _currentContext.SetupGet(x => x.OrganizationFeatures).Returns(organizationFeatures);
-        var features = _fixture
-            .Build<ApplicationFeatureContext>()
-            .With(x => x.EventLoggingIsEnabled, true)
-            .Create();
-        _currentContext.SetupGet(x => x.Features).Returns(features);
-
-        var organization = _fixture.Build<Organization>()
-            .Without(x => x.Applications)
-            .Without(x => x.Admins)
-            .Create();
-        _currentContext.SetupGet(x => x.Organization).Returns(organization);
-        _currentContext.SetupGet(x => x.IsPendingDelete).Returns(false);
-        _currentContext.SetupGet(x => x.InAppContext).Returns(true);
-        _authorizationContext.SetAuthorized("jonas");
-
-        // Act
-        var cut = RenderComponent<NavMenu>();
-
-        // Assert
-        var submenu = cut.Find("div[id=\"app-submenu\"]");
-        var links = submenu.Children.Where(x => x.ClassName == "nav-link").ToImmutableList();
-        Assert.Equal(6, links.Count);
-        Assert.Contains("Getting Started", links[0].TextContent);
-        Assert.Contains("Playground", links[1].TextContent);
-        Assert.Contains("Users", links[2].TextContent);
-        Assert.Contains("Reporting", links[3].TextContent);
-        Assert.Contains("Settings", links[4].TextContent);
-        Assert.Contains("App Logs", links[5].TextContent);
-    }
 }
